Remove generated enum; use const definitions only.
Since we want the type to be 'open' to values not known in the spec, an
enum seems a bad fit because of problems due to upgrades adding new
variants that might have been used by an application using the 'Unknown'
variant in an earlier release.  Since I can't see a way around that
upgrade-problem, the enum goes.